widget-examples/ssrm_mode/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
fastapi-cache2==0.2.1
python-multipart==0.0.6
mysql-connector-python==8.2.0
requests==2.31.0